+++ src/lib/checkPermissions.js
@@ -1,0 +1,8 @@
+export function checkPermissions(action) {
+	chrome.permissions.contains(
+		{
+			origins: ['<all_urls>'],
+		},
+		(result) => action(result)
+	);
+}
