+++ src/lib/createSettingsEditor.js
@@ -1,0 +1,13 @@
+import { useSettings } from './useSettings';
+
+export function createSettingsEditor() {
+	const editorContainer = document.getElementById('JSONeditor');
+	const settingsEditor = new window.JSONEditor(editorContainer, {
+		mode: 'code',
+	});
+
+	useSettings((settings) => settingsEditor.set(settings));
+
+	return settingsEditor;
+}
+
