+++ src/lib/defaultSettings.js
@@ -1,5 +1,4 @@
-// Default settings for first-time user
-const defaultSettings = {
+export const defaultSettings = {
 	style: 'font-size: 18px; font-family: monospace;',
 	markers: [
 		{
@@ -40,14 +39,3 @@ const defaultSettings = {
 	],
 };
 
-// Check storage for saved user settings
-export function useSettings(action) {
-	chrome.storage.sync.get(['settings'], function (result) {
-		try {
-			const settings = JSON.parse(result.settings);
-			action(settings);
-		} catch {
-			action(defaultSettings);
-		}
-	});
-}
