+++ src/lib/prependDocumentTitle.js
@@ -1,0 +1,4 @@
+export function prependDocumentTitle(name) {
+	document.title = `${name} — ${document.title}`;
+}
+
