+++ src/lib/saveSettings.js
@@ -1,0 +1,17 @@
+export function saveSettings(settingsEditor) {
+	const saveButton = document.getElementById('saveButton');
+
+	saveButton.addEventListener('click', () => {
+		try {
+			const settingsValue = settingsEditor.get();
+
+			alert('😁 OK! Reload pages to see your changes');
+			chrome.storage.sync.set({
+				settings: JSON.stringify(settingsValue),
+			});
+		} catch {
+			alert('🤬 Not valid JSON! Check your code...');
+		}
+	});
+}
+
