+++ src/lib/useSettings.js
@@ -1,0 +1,10 @@
+export function useSettings(action) {
+	chrome.storage.sync.get(['settings'], function (result) {
+		try {
+			const settings = JSON.parse(result.settings);
+			action(settings);
+		} catch {
+			action(defaultSettings);
+		}
+	});
+}
