+++ src/settings.html
@@ -78,8 +78,9 @@
 			<section>
 				<div id="JSONeditor"></div>
 
-				<button id="settingsButton" type="button">Save</button>
+				<button id="saveButton" type="button">Save</button>
 			</section>
 		</form>
 	</body>
 </html>
+
